[LETS-34] Deactivate record analysis/undo/finish postpone in page server recovery (#2847)
http://jira.cubrid.org/browse/LETS-34

Recovery on the page server must only apply unflushed change in the data pages. It must not generate any log records and it does not have to recovery the state of transaction table. As consequence, the next operations are disabled:

  - Almost all log record analysis. Log record analysis is usually run to rebuild the transaction table state before the server died or to find the end point for partial recovery. The only analysis required on the page server is on the last log record (end of log) to reset append LSA.
  - No undo of system operations, no finishing postpones and no undo of transactions.

Another necessary change is to disable replication verification in logpb_checkpoint, if it is run during recovery. The replicator is not initialized yet.