[LETS-118]alterdbhost does not mount/dismount active log (#2852)
alterdbhost changes the hostname set in databases.txt.

The implementation also mounts/dismounts the active log file, probably as a check that the database really exists. It is not mandatory for the mission of alterdbhost utilitary and it won't work with transaction servers running on remote storage (and it won't work with any transaction servers in the end).

Changed the code for alterdbhost to not mount/unmount the log active file.
The functionality of alterdbhost should not be altered by the change.