[LETS-149] Treat RVBT_MARK_DELETED as MVCC log record (#2863)
http://jira.cubrid.org/browse/LETS-149

RVBT_MARK_DELETED is a log record index used for deletes in the serial name index. Although the updates on the serials does not abide the MVCC rules, an MVCCID is used on deletes to protect against unique constraint violation in the next case:

- T1 deletes serial s1 without committing
- T2 inserts serial s1
- T1 aborts and wants to reinstate s1
- Unique constraint is violated.

However, RVBT_MARK_DELETED is not used as a log record for an MVCC operation. In consequence, when it is replicated, the page server does not know to advance `log_Gl.hdr.mvcc_next_id` and `btree_check_valid_record()` fails after the update on the serial name index (it expected all encountered MVCCID's are less than `log_Gl.hdr.mvcc_next_id`).

This patch includes RVBT_MARK_DELETED to the list of MVCC log record indexes. Vacuum is also instructed to skip it, because the record MVCCID is cleared by the transaction and no cleanup is required.