[LETS-59] Recovery speedup by caching log pages in log page buffer (#2879)
http://jira.cubrid.org/browse/LETS-59

Function `logpb_copy_page` function first checks if the page is in log page buffer (`pageid` must match with the corresponding slot). If this does not match, then the page is read from disk (or, in the scalability setup, from the page server).
After reading the page from storage (disk or page server), logic attempts to save the page in the log page buffer using a "highest page id" criteria.