[LETS-135] Transaction server thinks it is connected to a page server that does not exist (#2883)
The communication setup protocol between TS and PS has a flaw. If cub_master on PS node works, but the page server doesn't, the transaction server wrongfully thinks it has a valid connection. The send_int that is supposed to confirm the connection does not actually confirm that the message is received.

Fixed the connection protocol by adding a confirmation message from the page server side. After the ATS sends cubcomm::server_server::CONNECT_ACTIVE_TRAN_TO_PAGE_SERVER to the page server it will wait to receive the same value as a confirmation. If the same value is not received the ATS-PS connection is not created.