[LETS-159] Add request server start/stop thread error logging (#2887)
http://jira.cubrid.org/browse/LETS-159

Add error log messages when the request server is started and stopped. The message include information like the object address and the thread id.

It was originally incorporated into LETS-148 to investigate an issue that didn't reproduce a second time. It should help future efforts to reproduce and investigate it.