[LETS-148] Replicate unique index statistics (#2889)
http://jira.cubrid.org/browse/LETS-148

Changing statistics in the b-tree root page on every operation is very expensive. The transaction server is optimized to update these statistics only in-memory and log its changes. The statistics are flushed to b-tree root page during checkpoint.

The redo recovery follows the same system. Statistics are updated only in memory and they are flushed during checkpoint (recovery always ends with a checkpoint).

On the page server however, the replication has to diverge from the redo recovery system. B-tree root pages may be fetched at any time and the statistics should be there. Therefore, the replication has to replicate statistics changes before the page is fetched.

The easiest option is to replicate the changes immediately, when the log record is found.

**Implementation**

- Add special handling for RVBT_LOG_GLOBAL_UNIQUE_STATS_COMMIT (replicator::read_and_redo_btree_stats). It gets btid & stats from redo data and applies the changes.
- Add redo_job_btree_stats to parallelize b-tree stats replication. The VPID is obtained from the btid read from redo data.
- Add replicate_btree_stats, to apply the stats change and set the new page LSA.

**Other:**

- Move redo_job_replication_delay_impl to log_replication.cpp
- Rewrite btree_reflect_global_unique_statistics. Remove unused only_active_tran argument. Extract the code that changes stats in the b-tree root header to btree_root_update_stats
- Expose log_rv_redo_fix_page for log_replication.cpp
- Function to read btid + stats from recovery data.