[LETS-174] Quick find tran_index for tranid during log_recovery_redo (#2936)
http://jira.cubrid.org/browse/LETS-174

When LOG_COMMIT/LOG_ABORT records are found during log_recovery_redo, the transaction descriptors with the same tranid are freed. The transaction descriptor is searched by tranid in the transaction table. When tranid is not found (very likely), the entire transaction table is iterated. Depending on the configuration, the transaction table may have many thousands entries.

A quick optimization is to avoid looking up when tranid's are not in the transaction table. In most cases, these tranid's preceed the tranid's in the transaction table. The fix is to save the smallest tranid in the transaction table and compare the committed tranid with this. If the committed tranid is smaller, then searching in the transaction table is no longer necessary.