[LETS-131] Adapt disk volume add/expand to transaction server with remote storage (#2935)
http://jira.cubrid.org/browse/LETS-131

Do not create/extend the physical file of a volume on the transaction server with remote storage disk. Disable all log and data flushing to disk.

Assume all volume ID's are consecutive when getting the ID of previous volume.

disk_is_page_sector_reserved_with_debug_crash does not look for volume descriptor on TSRS. It was also rewritten with a unified exit routine.