[LETS-188] Transmit server type when server connection with master is re-established (#2980)
http://jira.cubrid.org/browse/LETS-188

If the connection with master is lost and then re-established, the server type is not transmitted. The first character in the server name is wrongfully interpreted as server type and the name gets trimmed.

Fix by always transmitting the server type first. css_pack_server_name was modified to append server type as suffix. The function name was changed to css_pack_message_to_master to be self-explanatory.

Other variables have been renamed to include message_to_master in the name.