[CBRD-23622] Truncate Table Performance Enhancements (merge) (#2861)
http://jira.cubrid.org/browse/CBRD-23622

TRUNCATE TABLE statement is used to delete all the data from an existing table but not the table itself. It has the same effect as DELETE FROM [table_name] semantically. 
In CUBRID, TRUNCATE is implemented internally using the DELETE FROM statement, which makes it scan and delete all the data record by record.
With this PR, it's done by deallocating pages that contain records.

It includes some details:
- Some specification changes and a new options, CASCADE.
- Still using DELETE when a table is refenced by an object reference.
- Some optimization releated to reacreation indexes during TRUNCATE.