[CBRD-24025] Speedup redo recovery by avoiding "is reserved" check (#2974)
https://jira.cubrid.org/browse/CBRD-24025

During recovery, there is a step that checks if page is "reserved". The check fixes sector reservation table page and searches for the bit corresponding to the page's sector. If it is not reserved, applying redo is skipped. As a result, if a file is destroyed before crash (drop index, drop table), many redo changes may be skipped. Skipping these redo record is not a strict requirement; it may have been an optimization and it may pass some automatic page buffer fix checks.

However, the performance impact of the check is serious. Measurements in some artificial tests showed an overall 2x improvement of the recovery redo part of the recovery sequence without the checks.

The check is completely avoidable by applying redo on pages regardless of its reserved status. In the worst case scenario, some redo changes may be applied unnecessarily before the page is unreserved. Considering the impact of the check though, it is likely to get better performance by applying the redo without doing any checks instead of doing the check and skipping the redo.

Implementation:
- add a new page fetch mode, for recovery, that accepts any page status: new page, de-allocated page, or normal page.
- use new page fetch mode during redo recovery/replication
- no "is page sector reserved" check (update log_rv_redo_page_fix)