[LETS-143] Add the type of server to the client-server connections (#3003)
Client need to specify the desired server typoe when making a connection. Cub_master already differentiates the servers between them by type, the change is to make the distinction when returning the actual connection.

All clients should be able to connect to the transaction server. Some client types should be able to connected to the page server too. These clients need to pass along to cub_master the desired type of server.

The clients that can connect to the page server too: backupdb, paramdump, statdump, csql with --sysadm.