[LETS-188] Fix server name after connection with master is re-established (#3015)
http://jira.cubrid.org/browse/LETS-188

The name registered on restarted cub_master after the server reconnects starts with the number character equivalent to the server type. The explanation is that css_Master_server_name global used to include the server type.

Fix by setting css_Master_server_name value to server name instead of the message sent to cub_master. Message packing was pushed down to css_init as a consequence.