[LETS-195] tran server remote storage recovery (#3057)
http://jira.cubrid.org/browse/LETS-195

Implement log recovery on a transaction server with remote storage:

After a crash, the transaction server running with remote storage (TSRS) is responsible with finishing all the transaction that were running at the time it crashed. Finishing transactions means running all postpones and undoing all uncommitted changes and requires appending more log records. Since page servers (PS) are not allowed to do this kind of operations, the responsibility lies with the transaction server.

TSRS needs to rebuild the transaction table image at the time of the crash to finalize the transactions. The image is reacreated starting with a snapshot of the transaction table saved by LETS-194, and running recovery analysis from the checkpoint LSA. Then transactions may be finalized.

Implementation:
- fork log_recovery into a new function - log_recovery_finish_transactions - for TSRS
- get the last transaction table checkpoint (snapshot) information from log meta
- recovery analysis starting with checkpoint LSA
- no recovery redo is performed
- abort all atomic system operations
- finish transactions postponed actions
- recovery undo
- set the prev_lsa (LSA of last log record) to function vacuum_notify_server_crashed; vacuum will start looking for MVCC log records older than this point
- clean up all system transactions descriptors
- initialize the locator
- re-start class representation cache

Other:
- fixed some compiler warnings
- harmonized some bool params