[LETS-240] Set clean shutdown metalog file on all servers (#3067)
http://jira.cubrid.org/browse/LETS-240

Metalog clean shutdown flag is used by a transaction server with remote storage to know whether it was crashed or normally shut down. For now, only transaction server with remote storage uses it.

However, when database is created, the flag is not set, because server is not considered a transaction server with remote storage. On the first boot of transaction with remote storage, recovery is run unnecessarily.

Fix by setting the clean shutdown flag in metalog regardless of server/storage type.