[CBRD-24073] Problem that predicate push does not work when there is a correlated constant value in another subquery. (#3056)
http://jira.cubrid.org/browse/CBRD-24073

The constant value of the select_list of the subquery is replaced with the equality term of the main query. Therefore, qo_reduce_equality_terms() must be called in post order.
```
select ...
 from (select col ==> 1 (1st replace)
          from ...
         where col = 1) a
 where a.col = b.col1 ==> 1 = b.col1 (2nd replace)
```
Only SELECT cases apply post order calls. For other cases, make changes later if necessary.
And the constant value of the select list is stored separately. This is used for substitution.