[LETS-242] Fix replication of heap records undo with vacuum cases (#3084)
http://jira.cubrid.org/browse/LETS-242

This is a fix of a replication issue in a very particular scenario. Sometimes when applying undo on heap records, the record is also vacuumed. However, the compensation log record does not reflect the vacuuming and the page server cannot and does not do the vacuum by itself.

The fix is to add a compensation log record that reflects the vacuumed heap log record.

**Details**

Intially, `vacuum_rv_check_at_undo` was updated to reflect all these changes, but it became an unmanageable monster. To make things more clear, the function was removed and the three functions - `heap_rv_undo_delete`, `heap_rv_undo_update` and `heap_rv_undo_ovf_update` have been rewritten.

A very detailed description of these three scenarios and why they need vacuuming and manual log compensation handling can be found at `heap_file.c:16214`.

`heap_rv_undo_ovf_update` is mostly the same as `vacuum_rv_check_at_undo` in the `REC_BIGONE` case.

There are some differences in `heap_rv_undo_delete` and `heap_rv_undo_update`. They used to apply the heap record change first, then read the modified record to verify if vacuum is required and modified the page again if it was. This pull requests cleans the record descriptor first and then applies the change into the heap page.

Both functions are very similar and the whole implementation can be found in `heap_rv_undo_with_vacuum_internal ()`. The only differences are:

  - Undo delete does not have to vacuum when the record type is REC_HOME. REC_HOME is never physically deleted during an MVCC heap update operation.
  - To undo the delete (and redo its compensation), `spage_insert_for_recovery` needs to be applied. To undo the update, `heap_update_physical` must be applied.

New recovery indexes have been added for new compensation cases: to compensate with an insert, with an update and with no operation.