[LETS-233] separate metalog for transaction server and page server (#3098)
http://jira.cubrid.org/browse/LETS-233

The single file that a TSRS actually uses/maintains is the metalog (for transaction table snapshoting). In order to allow execution of TSRS and PS on a single node, the servers needs to access separate metalog files. This should also cover the need of the recovery process on TSRS.

Implementation:
- a TSRS would create a metalog file upon boot-up, if not already found on disk
- separate name for TSRS-specific metalog file: _lgmeta_tran