[LETS-275] writing to meta log is fatal only upon initialize (#3146)
http://jira.cubrid.org/browse/LETS-275

Writing to metalog is done for the following purposes:
 - saving transaction table snapshot
 - registering the clean shutdown flag

Writing to metalog is considered a fatal error in all cases, as of now.

Some of the automated tests are flawed in the sense that, at the end, the script removes the database files before actually closing the server. With the introduction of the metalog, for the uses described above, writing the metalog causes many automated tests to cause core dumps.

Fix the scripted automated tests by only considering fatal the access of metalog upon initialization. All other writes are to be reported as non-fatal errors.