[LETS-276] Allow ATS reconnection after a crash (#3153)
http://jira.cubrid.org/browse/LETS-276

When ATS connects to the PS, there is an assert on the PS side that no prior ATS connection exists. During a normal shutdown, the ATS will disconnect from the PS, but after a crash the connection remains. When the ATS reconnects, the PS crashes.

Fix by accepting the fact that a connection with the ATS may already exist when a new connection is created. Delete the prior connection.