[CBRD-24069] Prevent emergency_patch -r if the active volume is sane enough (#3132)
http://jira.cubrid.org/browse/CBRD-24069

The emergency_patchlog utility provides the option, -r, to recreate a log for when the log is unreliable or inconsistent. However, this option tends to be overused even though the active log is sane. So, with this option, the utility will check whether the active log is reliable enough and denies it (failure) if it is. 

The criteria of reliability of an active log volume:
- Pass availability check in the given path
- Pass the compatibility check
- Pass the checksum check of the log header page