[LETS-187] Create database with remote storage (#3182)
http://jira.cubrid.org/browse/LETS-187

Create a database that relies on remote storage. The database entry is added to databases.txt file and paths configuration is handled similarly to a database with local storage.

Creating a database with remote storage may not overwrite an existing database. The createdb command throws an error when both -r and --remote-storage options are used.

Changes:

- Moved all path configuration initialization to `boot_initialize_paths`
- Added `--remote-storage` createdb option
- The `--remote-storage` version of createdb does `boot_initialize_paths` and `boot_initialize_remote_storage_client`