[LETS-288] Always log assigned MVCCID before completing transaction (#3192)
http://jira.cubrid.org/browse/LETS-288

All MVCCID's assigned on ATS are also completed when the transaction is committed or aborted. The MVCCID's have to completed in the same way on the PTS.

The only way PTS may be notified of assigned/completed MVCCID's is through the log. There are MVCC log records that include the transaction MVCCID. However, it is not guaranteed that all assigned MVCCID's end up being logged in an MVCC log record.

If an assigned MVCCID is not logged before its completion through an MVCC log record, it must be logged there.

- Track the last LSA of transaction MVCC log record.
- Before completing MVCC, if the transaction is assigned an MVCCID but it its last MVCC lsa is null, append a new type of log record containing the assigned MVCCID.

The same mechanism needs to be applied when transactions are aborted during recovery undo. Therefore, last MVCC lsa was included in checkpoint/analysis.

Another patch will have to do the MVCCID assignment/completion during PTS replication.