[CBRD-24102] Applylogdb may skip creating TDE table with the TDE module unloaded (#3168)
http://jira.cubrid.org/browse/CBRD-24102

In HA, if the slave node can't load the TDE module and runs into a statement related to TDE, it is supposed to just stop. However, it skips a CREATE TABLE ENCRYPT statement after about 10 minutes. 
Cause: During failure to create table due to the not loaded TDE-module, it actually creates a table (file), fails, and destroys it. These files are registered as dropped files. Once a file is registered as a dropped file, even if the file header page is deallocated, it cannot be used as a header page of another newly created file by a transaction of which mvccid is lower than the mvccid of the dropped file. The mvccid of dropped file is the 1 + mvcc id of the transaction destroying the file.
Solution: Moving the blocking point to do_create_entity() in the case of "CREATE TABLE".