[LETS-326] Rename request_sync_client_server::sequenced_payload (#3232)
http://jira.cubrid.org/browse/LETS-326

The name `request_sync_client_server::internal_payload` was not great, because the name "internal" would suggest that the class is not visible to the world outside `request_sync_client_server`. However, the handlers that have to be provided from outside use it in their argument signature.

Renamed as `sequenced_payload` because it adds the response sequence number to the regular payload that the requester wants to send to the other side.