[CBRD-23986] Revise the truncate query for dont_reuse_oid class (#3382)
http://jira.cubrid.org/browse/CBRD-23986

When doing TRUNCATE a class with the DONT_REUSE_OID option, a select query is used to determine the way to truncate it:
SELECT COUNT(*) FROM [_db_domain] WHERE [data_type]=%d AND ([class_of].[class_name]='%s' OR [class_of] IS NULL) LIMIT %d

We add an index for consistent running time (#3376), but a query rewriter makes a query doing an additional sequential scan even if there is an index. To prevent it, we revise the query which has the same meaning. See the Jira issue comment for more information.