[CBRD-23959] Socket pooling to manage connection with Java SP Server on DB Server (#3381)
http://jira.cubrid.org/browse/CBRD-23959

The method executing module gets one from the socket object pool when starting Java SP. When the module is ended, the socket returns to the pool.

The number of sockets in the pool is maintained as much as the number of workers (according to the max_client parameter). If more sockets are needed, they are created temporarily and are not managed in the socket pool.