[CBRD-24217] Error in data communication protocol between Driver and CAS (#3396)
http://jira.cubrid.org/browse/CBRD-24217

In the FETCH phase of the data communication protocol between the driver and the CAS, the data size and data are put into the network buffer. At this time, the data structure according to the data size and data type must match.
In case of SQL_TINYINT type, data size should be 2 bytes, and data type should be short.