[CBRD-24257] It prevents the CAS memory size from increasing when a large amount of data is fetched through a query. (#3456)
http://jira.cubrid.org/browse/CBRD-24257

When a large amount of data was imported with a cursor type other than SQL_CURSOR_FORWARD_ONLY among the ODBC cursor types, the CAS (CGW) memory size increased.
It was confirmed that the memory size increases in SQLExecute for MySQL and SQLFetchScroll for Oracle.
When importing large amounts of data, if the memory size of CAS (CGW) exceeds the set memory size, CAS (CGW) restarts. At this time, the driver cannot get the data and an error occurs.