[LETS-310] Handle page desynchronization during btree_range_scan OID selection (#3388)
http://jira.cubrid.org/browse/LETS-310

`btree_range_scan_select_visible_oids` function selects all current key OID's. If the key has overflow OID's, the selection may be interrupted if the OID buffer is full.
Additional to this, on PTS, the fixing of the overflow OID page may return page desynchronization error. The case must be treated similarly with the full buffer issue.
The existing mechanism that interrupts the key processing and resumes on next `btree_range_scan` iteration relies on the fact that it won't interrupt immediately after processing the leaf record; it always interrupts after an overflow page.
On PTS, the page desync error may occur on the first overflow page too. Therefore, on PTS things need to be done a little differently to apply the same kind of mechanism:
- When a key is processed the first time, and it has an overflow page, fix the overflow page before processing the leaf record objects. If page desynchronization error occurs, interrupt `btree_range_scan_select_visible_oids`. If overflow page is fixed, keep it fixed.
- In the while that fixes overflow pages, fix overflow page only if it is not fixed already. On first iteration, the overflow page would have already been fixed, on subsequent iterations, not.
- In the while that fixed overflow pages, if page desynchronization occurs, set `bts->is_key_partially_processed` to `true` and save `bts->O_vpid` on the `last_visible_overflow`.
- `btree_range_scan` calls `btree_range_scan_handle_error` -> `btree_range_scan_handle_page_ahead_repl_error` after `btree_range_scan_select_visible_oids` and if `bts.repl_sync_lsa` is set, the scan iteration is interrupted.