[LETS-422] Atomic replication helper class (#3484)
A new helper class that is able to hold log records information for all log records that pertain to an atomic redo sequence - which is, in itself, part of a single transaction - and apply them.

    Implementation:

Created 2 new classes atomic_replication_helper and atomic_replication_unit. The atomic_replication_unit holds the information about one log record and exposes methods to fix and unfix the page and apply the redo. The unit class is a private internal class of the atomic_replication_helper, the helper class holds the map of TRANID to atomic_replication_unit.

The classes will be used in the new replicator flavor on the passive transaction server in the upcoming issues.