[CBRD-23845] Support SYNONYM (#3541)
http://jira.cubrid.org/browse/CBRD-23845

A table that is not owned by itself can be used only by using the name as owner_name.table_name. It is difficult to use a table by checking the owner name every time. So it supports synonyms. Synonym can be used as an alias for the name owner_name.table_name.