[CBRD-23846] [regression] Fix unexpected "SELECT authorization failure" (#3544)
http://jira.cubrid.org/browse/CBRD-23846

**Purpose**
Some regression tests failed caused by "SELECT authorization failure". I found some code that authentication check for java stored procedure is not disabled.

**Implementation**
Owner of stored procedure should not be checked. When getting MOP of a stored procedure, the authentication check was disabled. But when getting any attribute from the MOP, SELECT authorization failure occurred.

**Remarks**
For the CALL statement, it occurs the error. But for the SELECT statement, the error does not occur.