[LETS-355] update log recovery state on PTS to allow temporary volume check to work (#3594)
http://jira.cubrid.org/browse/LETS-355

Passive Transaction Server is allowed to create temporary volumes. The check whether a temporary volume is influenced, for better or for worse, by the recovery state global flag (`log_Gl.rcv_phase`). On PTS, this state must be - gratuitously, because it is has no real usage otherwise - updated at some point in order to allow the temporary volume check to work correctly.