[CBRD-24350] Fix the core dump due to allocating less memory than needed in CDC (#3605)
http://jira.cubrid.org/browse/CBRD-24350

Purpose

When allocating memory for log item, it calculate the length of the record by recdes->length.

If the record (string) is compressed, the exact length of the record cannot be known through recdes->length. So, I added a function to know the length of the attributes according to the data type.