[CBRD-24344] [Regression] Core dumped in cubbase::restrack_assert (#3609)
http://jira.cubrid.org/browse/CBRD-24344

The dblink_scan_open and dblink_scan_next looks not to handle error properly.

      ret = cci_execute (scan_info->stmt_handle, 0, 0, &err_buf);
      if (ret < 0)
        {
          er_set (ER_ERROR_SEVERITY, ARG_FILE_LINE, ER_DBLINK, 1, err_buf.err_msg);
          return ER_DBLINK; 
        }
The above code snipet shows not to close dblink scan at error case. If the dblink scan does not close, it might araise memry leak because the cci memory is not free while query execution.

      ret = cci_execute (scan_info->stmt_handle, 0, 0, &err_buf);
      if (ret < 0)
        {
          er_set (ER_ERROR_SEVERITY, ARG_FILE_LINE, ER_DBLINK, 1, err_buf.err_msg);
          dblink_scan_close (scan_info);
          return ER_DBLINK; 
        } 
So, I modified the code as above to close dblink_scan.