[LETS-289] new mvcc log record field - parent_mvccid (#3644)
http://jira.cubrid.org/browse/LETS-289

A subtransaction can have it's own mvccid. If that is the case, the transaction will also have a "main" mvccid (this is visible, among others, in the way the function `logtb_get_current_mvccid` is implemented).

With existing information it is not possible to know across replication boundaries (ie: on either page server or passive transaction server) whether the mvccid transmitted in log records is either a "main" mvccid or sub-mvccid.

Added a new `parent_mvccid` in the MVCC-related union member of `LOG_REC_SYSOP_END`.
If transaction has an mvccid allocated by a sub-transaction, this new field will contain the transaction's "main" mvccid (which, as indicated in implementations in `logtb_get_new_subtransaction_mvccid` and `logtb_get_current_mvccid`, must be valid).
Otherwise, the new field should be ignored.

The purpose of this field is twofold:
- to discerne the nature of the mvccid across replication on passive transaction server and also - as either 'main' mvccid or sub-mvccid
- and to allow completion of both id's during replication

Implementation:
- introduced new anonymous struct in `LOG_REC_SYSOP_END` as union member for mvcc and  added field `parent_mvccid`
- adapted use all over
- only populated the new field where relevant in `log_sysop_end_logical_undo` with newly added function `logtb_get_current_mvccid_and_parent_mvccid`
- logged new field in dump functions where relevant

Only populating the new field is present in this patch. Using it across replication boundary will be a separate patch.