[LETS-460] [Re]organize page fix, apply and page unfix (#3643)
Applying the log right after the fix could lead to problems as the records are fixed one by one as they come to be read by the PTS and some might be unfixed and refixed after the apply procedure leading to inconsistency.

To solve this a new order of operations has been added, the pages are now fixed one by one as necessary and then when the atomic sequence end log record appears all the records in the sequence are applied sequentially and then the pages are all unfixed. To provide the required context information for the apply the atomic_replication_sequence now has a new member in the form of the log_rv_redo_context m_redo_context and the logic follows the model of the replicatior class.