[LETS-423] Separate log replicator for page server and passive transaction server - Integrate the atomic_replicator_helper class (#3570)
In order for the atomic replication to work on PTS a new type replicator has to be created. The new atomic_replicator will replace the replicator for PTS.

The integration of the atomic_replication_helper is the first step towards the functioning new replicator.
The following PRs will differentiate the replicator types further.