[CBRD-24381] Add system parameter to set window size for QA test (#3713)
http://jira.cubrid.org/browse/CBRD-24381

Purpose

For QA test for flashback, it need to provide a parameter to set a window size.
When performing flashback, after measuring the window size, a summary is output according to the size.
However, if flashback is performed with the shell, the window size is recognized as 0 and proper testing cannot be performed.