[LETS-484] start pts with more than one data volume (#3737)
http://jira.cubrid.org/browse/LETS-484

On transaction server with remote storage, there is a separate bookkeeping of the number of volumes (which are actually created remotely, on the page server) that must be maintained.
On such a server, the check for volume id validity is done against this separate bookkeeping - `disk_is_valid_volid`.
This patch only maintains the separate bookkeeping for active transaction server. For passive transaction server, a separate patch will do the update of the value upon replicating the `RVDK` log records.