[LETS-507] reset all mvccids in the transaction table after analysis (#3776)
http://jira.cubrid.org/browse/LETS-507

During analysis, the mvccids are loaded from log records into the created transaction descriptors.
These mvccids are needed for the specific needs of initializing a passive transaction server (see function `log_recovery_analysis_from_trantable_snapshot`) with the final purpose of constructing a valid&up-to-date MVCC table (see function `log_recovery_build_mvcc_table_from_trantable`).
As a side effect, this also happens in all other cases, where that MVCCID is not needed (eg: recovery of  monolithic server). When analysis finishes, some of those transactions - those that have not yet committed and will be aborted anyway - still have those valid mvccids set. When such transactions descriptors will be cleaned, there is a sanity-check that the mvccid must be null.
Since the mvccid is only needed for the analysis step of the passive transaction server, for all other cases we clean-up those mvccids after the analysis step, before redo.

This is a fix for a regression introduced in tests pertaining to the monolithic server.

Implementation:
- add a new function to do this clean-up and call it after the `log_recovery_analysis` call