[LETS-511] Atomic replication for updating a heap record and its prev_version_lsa (#3787)
http://jira.cubrid.org/browse/LETS-511

Add the redo-only log record, RVHF_SET_PREV_VERSION_LSA, and append it with the atomic replication mechanism. Some heap operations already use the mechanism, so I add it in case it is not used.

See the Jira issue for more information about the issue and approach.

Log records to be applied atomically for each update case:

REC_HOME -> * : For any case, it's enough to set prev_version_lsa and update the home record atomically.
REC_RELOCATION:
-> REC_HOME, REC_BIGONE, and REC_RELOCATION on the same page: The home record is updated, and it has to be applied atomically with setting prev version lsa. Actually, for all cases, LOG_START_ATOMIC_REPL is already appended, but I have added it conditionally for safety.
-> REC_RELOCATION on another page: The home record is not updated, so the forward record has to be updated atomically with setting prev version lsa.
REC_BIGONE -> * : The BIGONE redo record already has the previous version lsa, so we don't need to care about this case.
Note. I have tested it with the REC_HOME records with the scenario in the Jira issue and see it works and gets a consistent result. However, I haven't tested with every heap log type transition such as REC_HOME -> REC_RELOCATION and REC_NEWHOME.